extract some common functions from spark to iceberg-core
