+++ spark/src/main/java/org/apache/iceberg/actions/ExpireSnapshotsAction.java
@@ -53,7 +53,7 @@ import org.slf4j.LoggerFactory;
  * are issued. Deletes are still performed locally after retrieving the results from the Spark executors.
  */
 @SuppressWarnings("UnnecessaryAnonymousClass")
-public class ExpireSnapshotsAction extends BaseAction<ExpireSnapshotsActionResult> {
+public class ExpireSnapshotsAction extends SparkBaseAction<ExpireSnapshotsActionResult> {
   private static final Logger LOG = LoggerFactory.getLogger(ExpireSnapshotsAction.class);
 
   private static final String DATA_FILE = "Data File";
