+++ spark/src/main/java/org/apache/iceberg/actions/SparkBaseAction.java
@@ -24,14 +24,11 @@ import java.util.List;
 import org.apache.iceberg.BaseTable;
 import org.apache.iceberg.ManifestFiles;
 import org.apache.iceberg.MetadataTableType;
-import org.apache.iceberg.Snapshot;
 import org.apache.iceberg.StaticTableOperations;
 import org.apache.iceberg.Table;
-import org.apache.iceberg.TableMetadata;
 import org.apache.iceberg.TableOperations;
 import org.apache.iceberg.io.ClosingIterator;
 import org.apache.iceberg.io.FileIO;
-import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.spark.SparkUtil;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.FlatMapFunction;
@@ -41,62 +38,7 @@ import org.apache.spark.sql.Encoders;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
 
-abstract class BaseAction<R> implements Action<R> {
-
-  protected abstract Table table();
-
-  protected String metadataTableName(MetadataTableType type) {
-    return metadataTableName(table().name(), type);
-  }
-
-  protected String metadataTableName(String tableName, MetadataTableType type) {
-    if (tableName.contains("/")) {
-      return tableName + "#" + type;
-    } else if (tableName.startsWith("hadoop.")) {
-      // for HadoopCatalog tables, use the table location to load the metadata table
-      // because IcebergCatalog uses HiveCatalog when the table is identified by name
-      return table().location() + "#" + type;
-    } else if (tableName.startsWith("hive.")) {
-      // HiveCatalog prepend a logical name which we need to drop for Spark 2.4
-      return tableName.replaceFirst("hive\\.", "") + "." + type;
-    } else {
-      return tableName + "." + type;
-    }
-  }
-
-  /**
-   * Returns all the path locations of all Manifest Lists for a given list of snapshots
-   * @param snapshots snapshots
-   * @return the paths of the Manifest Lists
-   */
-  private List<String> getManifestListPaths(Iterable<Snapshot> snapshots) {
-    List<String> manifestLists = Lists.newArrayList();
-    for (Snapshot snapshot : snapshots) {
-      String manifestListLocation = snapshot.manifestListLocation();
-      if (manifestListLocation != null) {
-        manifestLists.add(manifestListLocation);
-      }
-    }
-    return manifestLists;
-  }
-
-  /**
-   * Returns all Metadata file paths which may not be in the current metadata. Specifically
-   * this includes "version-hint" files as well as entries in metadata.previousFiles.
-   * @param ops TableOperations for the table we will be getting paths from
-   * @return a list of paths to metadata files
-   */
-  private List<String> getOtherMetadataFilePaths(TableOperations ops) {
-    List<String> otherMetadataFiles = Lists.newArrayList();
-    otherMetadataFiles.add(ops.metadataFileLocation("version-hint.text"));
-
-    TableMetadata metadata = ops.current();
-    otherMetadataFiles.add(metadata.metadataFileLocation());
-    for (TableMetadata.MetadataLogEntry previousMetadataFile : metadata.previousFiles()) {
-      otherMetadataFiles.add(previousMetadataFile.file());
-    }
-    return otherMetadataFiles;
-  }
+abstract class SparkBaseAction<R> extends BaseAction<R> implements Action<R> {
 
   protected Dataset<Row> buildValidDataFileDF(SparkSession spark) {
     return buildValidDataFileDF(spark, table().name());
